Breadcrumbs and PageHeaderTitle fix
Fixes use in pages with dynamic values in the pathname. usePathname() doesn't return the dynamic values in the pathname, so a util function to get them through useParams() was added. Util function "getPathnameArrayWithParamsKeys()" recieves the pathname and the params and returns an array of strings with each values in the pathname including the dynamic values.